Bugfix - cannot get cluster info in node version < 6 (#153)
The cluser.on('message') function doesn't have the worker as the first
argument in node < 6. To do this, the official way is to check the number of
arguments received. However, because this is using a fat arrow function (() => {}),
the variable `arguments` is hoisted outside the scope so the actual arguments
has a length of 5.

This is best demonstrated by viewing this Babel snippet - http://bit.ly/2wbApcb