fix: avoid mutation bug in registry
The defaultLabels feature is buggy when used in combination
with the `.labels()` sub counter feature.

We mutate the object in the serialization implementation which
means the next time you observe the value the labels object has
been incorrectly mutated.

The functionality of `.labels()` is supposed to be the caching
of the labels object itself instead of constantly re-using it.