Rework hashObject to optimize for label count > 1
The code as written has a bunch of bailouts for zero or one labels.

The benchmarks that the no-label code path is over 500 times faster
than the label codepath, which is the median path that users will
actually experience.

This version collapsed the bailouts and assumes that we are going
to see multiple labels. This leave no labels about even but
around 6-10% gains for the common path. Which is now only 440x
slower.

Side note: I made multiple attempts at switching this to [].join(','),
which would also shorten all of the keys by one, but the performance
was just not there. string += interpolated string is just a little
faster than the cost of the temporary array for the join().